services/ml/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pgvector==0.2.5

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication (for JWT validation)
python-jose[cryptography]==3.3.0

# Redis
redis==5.0.1

# HTTP client (for scraping)
httpx==0.26.0

# NLP and ML
numpy==1.26.4
sentence-transformers==2.5.1
trafilatura==1.8.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
aiosqlite==0.19.0
